Remove the RCM singleton, move more responsibility to the subscription manager
